# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "lksh-enter-2025",
    "summary": "Test assignment for lksh-2025",
    "version": "0.1.0"
  },
  "paths": {
    "/stats": {
      "get": {
        "summary": "Get team's statistics",
        "operationId": "get_team_stats_stats_get",
        "parameters": [
          {
            "name": "team_name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Team Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/STeamStatsResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "example": {
                  "detail": [
                    {
                      "loc": [
                        "string",
                        0
                      ],
                      "msg": "Team not found",
                      "type": "string"
                    }
                  ]
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/versus": {
      "get": {
        "summary": "Get players' games' amount",
        "operationId": "get_player_versus_versus_get",
        "parameters": [
          {
            "name": "player1_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Player1 Id"
            }
          },
          {
            "name": "player2_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Player2 Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SVersusResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "example": {
                  "detail": [
                    {
                      "loc": [
                        "string",
                        0
                      ],
                      "msg": "Team is not found",
                      "type": "string"
                    },
                    {
                      "loc": [
                        "string",
                        0
                      ],
                      "msg": "Player is not found",
                      "type": "string"
                    }
                  ]
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/goals": {
      "get": {
        "summary": "Get player's goals",
        "operationId": "get_goals_goals_get",
        "parameters": [
          {
            "name": "player_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Player Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SGoalsRequest"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "SGoal": {
        "properties": {
          "match": {
            "type": "integer",
            "title": "Match"
          },
          "time": {
            "type": "integer",
            "title": "Time"
          }
        },
        "type": "object",
        "required": [
          "match",
          "time"
        ],
        "title": "SGoal"
      },
      "SGoalsRequest": {
        "properties": {
          "player_id": {
            "type": "integer",
            "title": "Player Id"
          }
        },
        "type": "object",
        "required": [
          "player_id"
        ],
        "title": "SGoalsRequest"
      },
      "SGoalsResponse": {
        "items": {
          "$ref": "#/components/schemas/SGoal"
        },
        "type": "array"
      },
      "STeamStatsResponse": {
        "properties": {
          "wins": {
            "type": "integer",
            "title": "Wins"
          },
          "losses": {
            "type": "integer",
            "title": "Losses"
          },
          "balance": {
            "type": "integer",
            "title": "Balance"
          }
        },
        "type": "object",
        "required": [
          "wins",
          "losses",
          "balance"
        ],
        "title": "STeamStatsResponse"
      },
      "SVersusResponse": {
        "properties": {
          "amount": {
            "type": "integer",
            "title": "Amount"
          }
        },
        "type": "object",
        "required": [
          "amount"
        ],
        "title": "SVersusResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}